pyproject.toml
[project]
name = "mysql"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=2.3.3",
    "pymysql>=1.1.1",
]
